Modify ElasticPress index and mapping
Creates an index name that is timestamped with the date. Adds in the
Elasticsearch response, which is much more detailed and helpful then the
ElasticPress response. Add in a new custom MoJ Analyzer that introduces
synonyms. Further commits will allow the user to add synonyms, right now
this commit just hardcodes the synonyms - more as a demonstration.