Add synonym checks, better documenation, aws packages
Synonyms are added via array merging, added checks for that process.
Further documentation in code and the addition of synonyms referencing
an AWS package rather then hardcoded synonyms. This is still somewhat
hardcoded however, as a change to your synonyms.txt will require
re-adding and indexing the reference path. This will be improved in
further commits, however, this is now fully working and being applied to
new indexies.