Proof of concept lambda to s3 ES index method
At a very basic level, this streams json files from our app to a lambda
function in AWS, to an s3 bucket. Then they are picked up from the s3
bucket in our app and sent off to ES. This commit works however it is
very slow and will require some refinement if we want to implement it.